requirements.txt
sqlalchemy==2.0.38
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0
